deployment/model-package/requirements/requirements.txt
numpy>=1.23.0,<3.0.0
pandas>=2.0.0,<4.0.0
scikit-learn>=1.3.2,<2.0.0
joblib>=1.1.0,<2.0.0
PyYAML>=6.0.0,<7.0.0
